requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
aiostream==0.4.4
altair==4.2.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
Babel==2.10.3
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
blobfile==1.3.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
commonmark==0.9.1
cryptography==37.0.4
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
devtools==0.8.0
discoart==0.7.11
docarray==0.13.33
docker==5.0.3
entrypoints==0.4
executing==0.8.3
fastapi==0.79.0
fastjsonschema==2.16.1
filelock==3.7.1
fonttools==4.34.4
frozenlist==1.3.0
ftfy==6.1.1
gitdb==4.0.9
GitPython==3.1.27
grpcio==1.47.0
grpcio-health-checking==1.47.0
grpcio-reflection==1.47.0
guided-diffusion-sdk==1.0.0
h11==0.13.0
huggingface-hub==0.8.1
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
jedi==0.18.1
jina==3.7.0
Jinja2==3.1.2
json5==0.9.8
jsonschema==4.7.2
kiwisolver==1.4.4
lpips==0.1.4
lz4==4.0.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.2
nest-asyncio==1.5.5
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.1
open-clip-torch==1.3.0
openai-clip==1.0.1
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==8.0.0
pycparser==2.21
pycryptodomex==3.15.0
pydantic==1.9.1
pydeck==0.7.1
Pygments==2.12.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
pyspellchecker==0.6.3
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==23.2.0
regex==2022.7.9
requests==2.28.1
resize-right-sdk==1.0.2
rich==12.5.1
scipy==1.8.1
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.3.0
starlette==0.19.1
streamlit==1.11.0
terminado==0.15.0
tinycss2==1.1.1
tokenizers==0.12.1
toml==0.10.2
toolz==0.12.0
torch==1.12.0+cu116
torchaudio==0.12.0+cu116
torchvision==0.13.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.20.1
typing-extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
uvicorn==0.18.2
uvloop==0.16.0
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
websockets==10.3
widgetsnbextension==3.6.1
xmltodict==0.12.0
yarl==1.7.2
zipp==3.8.1
